backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
redis==5.0.1
python-socketio==5.3.0
python-dotenv==1.0.0